restructure code from last review
